+++ settings.py
@@ -5,3 +5,6 @@
 
 # Time (in seconds) to wait before setting a link as up
 LINK_UP_TIMER = 10
+
+# Base URL of the Flow Manager endpoint
+FLOW_MANAGER_URL = "http://localhost:8181/api/kytos/flow_manager/v2"