+++ tests/unit/test_main.py
@@ -4,8 +4,8 @@
 import pytest
 import time
 from datetime import timedelta
-from unittest.mock import MagicMock, create_autospec, patch, call
-
+from unittest.mock import MagicMock, create_autospec, patch, call, Mock
+import tenacity
 from kytos.core.common import EntityStatus
 from kytos.core.helpers import now
 from kytos.core.events import KytosEvent
@@ -594,18 +594,28 @@ class TestMain:
         assert response.status_code == 404
         assert mock_switch.enable.call_count == 0
 
+    @patch('napps.kytos.topology.main.Main.notify_link_enabled_state')
     @patch('napps.kytos.topology.main.Main.notify_switch_links_status')
     @patch('napps.kytos.topology.main.Main.notify_topology_update')
-    async def test_disable_switch(self, mock_notify_topo, mock_sw_l_status):
+    async def test_disable_switch(self, *args):
         """Test disable_switch."""
+        mock_notify_topo, mock_sw_l_status, mock_noti_link = args
         dpid = "00:00:00:00:00:00:00:01"
         mock_switch = get_switch_mock(dpid)
+        interface = Mock()
+        interface.link.is_enabled = lambda: True
+        mock_switch.interfaces = {1: interface}
         self.napp.controller.switches = {dpid: mock_switch}
 
         endpoint = f"{self.base_endpoint}/switches/{dpid}/disable"
         response = await self.api_client.post(endpoint)
         assert response.status_code == 201
         assert mock_switch.disable.call_count == 1
+        assert mock_noti_link.call_count == 1
+        assert mock_noti_link.call_args[0][0] == interface.link
+        assert mock_noti_link.call_args[0][1] == "disabled"
+        assert interface.link.disable.call_count == 1
+        assert self.napp.topo_controller.bulk_disable_links.call_count == 1
         self.napp.topo_controller.disable_switch.assert_called_once_with(dpid)
         mock_notify_topo.assert_called()
         mock_sw_l_status.assert_called()
@@ -698,7 +708,13 @@ class TestMain:
             dpid, key
         )
 
-        # fail case
+        # 404, Metadata not found
+        mock_switch.metadata = {}
+        endpoint = f"{self.base_endpoint}/switches/{dpid}/metadata/{key}"
+        response = await self.api_client.delete(endpoint)
+        assert response.status_code == 404
+
+        # 404, Switch not found
         key = "A"
         dpid = "00:00:00:00:00:00:00:02"
         endpoint = f"{self.base_endpoint}/switches/{dpid}/metadata/{key}"
@@ -706,18 +722,35 @@ class TestMain:
         assert mock_metadata_changes.call_count == 1
         assert response.status_code == 404
 
+    # pylint: disable=too-many-statements
     @patch('napps.kytos.topology.main.Main.notify_topology_update')
     async def test_enable_interfaces(self, mock_notify_topo):
         """Test enable_interfaces."""
         dpid = '00:00:00:00:00:00:00:01'
         mock_switch = get_switch_mock(dpid)
         mock_interface_1 = get_interface_mock('s1-eth1', 1, mock_switch)
+        mock_interface_1.link = Mock()
+        mock_interface_1.link._enabled = True
         mock_interface_2 = get_interface_mock('s1-eth2', 2, mock_switch)
+        mock_interface_2.link = Mock()
+        mock_interface_2.link._enabled = False
         mock_switch.interfaces = {1: mock_interface_1, 2: mock_interface_2}
-        self.napp.controller.switches = {dpid: mock_switch}
 
+        # Switch not found
         interface_id = '00:00:00:00:00:00:00:01:1'
+        endpoint = f"{self.base_endpoint}/interfaces/{interface_id}/enable"
+        response = await self.api_client.post(endpoint)
+        assert response.status_code == 404
 
+        # Switch not enabled
+        mock_switch.is_enabled = lambda: False
+        self.napp.controller.switches = {dpid: mock_switch}
+        endpoint = f"{self.base_endpoint}/interfaces/{interface_id}/enable"
+        response = await self.api_client.post(endpoint)
+        assert response.status_code == 409
+
+        # Success
+        mock_switch.is_enabled = lambda: True
         endpoint = f"{self.base_endpoint}/interfaces/{interface_id}/enable"
         response = await self.api_client.post(endpoint)
         assert response.status_code == 200
@@ -757,14 +790,19 @@ class TestMain:
         assert mock_interface_1.enable.call_count == 0
         assert mock_interface_2.enable.call_count == 0
 
+    @patch('napps.kytos.topology.main.Main.notify_link_enabled_state')
     @patch('napps.kytos.topology.main.Main.notify_topology_update')
-    async def test_disable_interfaces(self, mock_notify_topo):
+    async def test_disable_interfaces(self, mock_notify_topo, mock_noti_link):
         """Test disable_interfaces."""
         interface_id = '00:00:00:00:00:00:00:01:1'
         dpid = '00:00:00:00:00:00:00:01'
         mock_switch = get_switch_mock(dpid)
         mock_interface_1 = get_interface_mock('s1-eth1', 1, mock_switch)
+        mock_interface_1.link = Mock()
+        mock_interface_1.link.is_enabled = lambda: True
         mock_interface_2 = get_interface_mock('s1-eth2', 2, mock_switch)
+        mock_interface_2.link = Mock()
+        mock_interface_2.link.is_enabled = lambda: False
         mock_switch.interfaces = {1: mock_interface_1, 2: mock_interface_2}
         self.napp.controller.switches = {dpid: mock_switch}
 
@@ -777,6 +815,11 @@ class TestMain:
         )
         assert mock_interface_1.disable.call_count == 1
         assert mock_interface_2.disable.call_count == 0
+        assert mock_interface_1.link.disable.call_count == 1
+        assert mock_noti_link.call_count == 1
+        assert mock_noti_link.call_args[0][0] == mock_interface_1.link
+        assert mock_noti_link.call_args[0][1] == "disabled"
+        assert self.napp.topo_controller.disable_interface.call_count == 1
         mock_notify_topo.assert_called()
 
         mock_interface_1.disable.call_count = 0
@@ -790,7 +833,14 @@ class TestMain:
             mock_switch.id, mock_switch.as_dict()
         )
         assert mock_interface_1.disable.call_count == 1
+        assert mock_interface_1.link.disable.call_count == 2
         assert mock_interface_2.disable.call_count == 1
+        assert mock_noti_link.call_count == 2
+        assert mock_noti_link.call_args[0][0] == mock_interface_1.link
+        assert mock_noti_link.call_args[0][1] == "disabled"
+        bulk_controller = self.napp.topo_controller.bulk_disable_links
+        assert bulk_controller.call_count == 1
+        assert len(bulk_controller.call_args[0][0]) == 1
 
         # test interface not found
         interface_id = '00:00:00:00:00:00:00:01:3'
@@ -921,39 +971,65 @@ class TestMain:
         response = await self.api_client.delete(url)
         assert response.status_code == 404
 
+    @patch('napps.kytos.topology.main.Main.notify_link_enabled_state')
     @patch('napps.kytos.topology.main.Main.notify_topology_update')
-    async def test_enable_link(self, mock_notify_topo):
+    async def test_enable_link(self, mock_notify_topo, mock_noti_link):
         """Test enable_link."""
         mock_link = MagicMock(Link)
+        link_id = "1"
+        mock_link.id = link_id
+        mock_link.is_enabled = lambda: False
+        mock_link.endpoint_a = MagicMock(is_enabled=lambda: True)
+        mock_link.endpoint_b = MagicMock(is_enabled=lambda: True)
         self.napp.links = {'1': mock_link}
 
-        link_id = "1"
+        # 409, endpoint is/are disabled
+        mock_link.endpoint_a.is_enabled = lambda: False
+        mock_link.endpoint_b.is_enabled = lambda: False
+        endpoint = f"{self.base_endpoint}/links/{link_id}/enable"
+        response = await self.api_client.post(endpoint)
+        assert response.status_code == 409
+
+        mock_link.endpoint_a.is_enabled = lambda: True
+        endpoint = f"{self.base_endpoint}/links/{link_id}/enable"
+        response = await self.api_client.post(endpoint)
+        assert response.status_code == 409
+
+        # Success
+        mock_link.endpoint_b.is_enabled = lambda: True
         endpoint = f"{self.base_endpoint}/links/{link_id}/enable"
         response = await self.api_client.post(endpoint)
         assert response.status_code == 201
-        assert mock_link.enable.call_count == 1
-        self.napp.topo_controller.enable_link.assert_called_with(link_id)
+        assert mock_noti_link.call_count == 1
+        assert mock_noti_link.call_args[0][0] == mock_link
+        assert mock_noti_link.call_args[0][1] == "enabled"
         mock_notify_topo.assert_called()
 
-        # fail case
+        # 404, link not found
         link_id = "2"
         endpoint = f"{self.base_endpoint}/links/{link_id}/enable"
         response = await self.api_client.post(endpoint)
         assert response.status_code == 404
+        assert mock_noti_link.call_count == 1
 
+    @patch('napps.kytos.topology.main.Main.notify_link_enabled_state')
     @patch('napps.kytos.topology.main.Main.notify_topology_update')
-    async def test_disable_link(self, mock_notify_topo):
+    async def test_disable_link(self, mock_notify_topo, mock_notify):
         """Test disable_link."""
         mock_link = MagicMock(Link)
+        mock_link.is_enabled = lambda: True
         self.napp.links = {'1': mock_link}
 
         link_id = "1"
         endpoint = f"{self.base_endpoint}/links/{link_id}/disable"
         response = await self.api_client.post(endpoint)
         assert response.status_code == 201
-        assert mock_link.disable.call_count == 1
         assert mock_notify_topo.call_count == 1
-        self.napp.topo_controller.disable_link.assert_called_with(link_id)
+        assert mock_notify.call_count == 1
+        assert mock_notify.call_args[0][0] == mock_link
+        assert mock_notify.call_args[0][1] == "disabled"
+        assert mock_link.disable.call_count == 1
+        assert self.napp.topo_controller.disable_link.call_count == 1
 
         # fail case
         link_id = "2"
@@ -1916,3 +1992,79 @@ class TestMain:
         assert response.status_code == 404
         assert self.napp.topo_controller.delete_link.call_count == 1
         assert self.napp.controller.buffers.app.put.call_count == call_count
+
+    @patch('napps.kytos.topology.main.Main.get_flows_by_switch')
+    async def test_delete_switch(self, mock_get):
+        """Test delete_switch"""
+        # Error 404 NotFound
+        dpid = '00:00:00:00:00:00:00:01'
+        endpoint = f"{self.base_endpoint}/switches/{dpid}"
+        response = await self.api_client.delete(endpoint)
+        assert response.status_code == 404, response
+
+        # Error 409 Switch not disabled
+        mock_switch = get_switch_mock(dpid)
+        mock_switch.status = EntityStatus.UP
+        self.napp.controller.switches = {dpid: mock_switch}
+        endpoint = f"{self.base_endpoint}/switches/{dpid}"
+        response = await self.api_client.delete(endpoint)
+        assert response.status_code == 409, response
+
+        # Error 409 Interface vlan is being used
+        mock_intf = MagicMock(all_tags_available=lambda: False)
+        mock_switch.interfaces = {1: mock_intf}
+        mock_switch.status = EntityStatus.DISABLED
+        endpoint = f"{self.base_endpoint}/switches/{dpid}"
+        response = await self.api_client.delete(endpoint)
+        assert response.status_code == 409, response
+
+        # Error 409 Swith have links
+        mock_switch.interfaces[1].all_tags_available = lambda: True
+        mock_switch_2 = get_switch_mock("00:00:00:00:00:00:00:02")
+        mock_interface_a = get_interface_mock('s1-eth1', 1, mock_switch)
+        mock_interface_b = get_interface_mock('s2-eth1', 1, mock_switch_2)
+        mock_link = get_link_mock(mock_interface_a, mock_interface_b)
+        self.napp.links = {'0e2b5d7bc858b9f38db11b69': mock_link}
+        endpoint = f"{self.base_endpoint}/switches/{dpid}"
+        response = await self.api_client.delete(endpoint)
+        assert response.status_code == 409, response
+
+        # Error 409 Switch has flows
+        mock_get.return_value = {dpid: {}}
+        endpoint = f"{self.base_endpoint}/switches/{dpid}"
+        response = await self.api_client.delete(endpoint)
+        assert response.status_code == 409, response
+
+        # Success 202
+        mock_get.return_value = {}
+        self.napp.links = {}
+        endpoint = f"{self.base_endpoint}/switches/{dpid}"
+        response = await self.api_client.delete(endpoint)
+        assert response.status_code == 200, response
+
+    def test_notify_link_status(self):
+        """Test notify_link_enabled_state"""
+        self.napp.controller.buffers.app.put.reset_mock()
+        link = Mock()
+        link.id = 'mock_link'
+        self.napp.notify_link_enabled_state(link, 'enabled')
+        assert self.napp.controller.buffers.app.put.call_count == 1
+
+        self.napp.notify_link_enabled_state(link, 'disabled')
+        assert self.napp.controller.buffers.app.put.call_count == 2
+
+    @patch('napps.kytos.topology.main.tenacity.nap.time')
+    @patch('httpx.get')
+    def test_get_flows_by_switch(self, mock_get, _):
+        """Test get_flows_by_switch"""
+        dpid = "00:01"
+        mock_get.return_value.status_code = 400
+        with pytest.raises(tenacity.RetryError):
+            self.napp.get_flows_by_switch(dpid)
+
+        mock_get.return_value.status_code = 200
+        mock_get.return_value.is_server_error = False
+        expected = {dpid: "mocked_flows"}
+        mock_get.return_value.json.return_value = expected
+        actual = self.napp.get_flows_by_switch(dpid)
+        assert actual == "mocked_flows"