+++ CHANGELOG.rst
@@ -18,7 +18,8 @@ Added
 - Added listener for ``kytos/core.interface_tags`` event to save any changes made to ``Interface`` attributes ``tag_ranges`` and ``available_tags``
 - Added script ``special_vlan_allocation.py`` to add ``special_available_tags`` and ``special_tags`` fields to ``interface_details`` collection.
 - Added endpoint ``POST v3/interfaces/{interface_id}/special_tags`` to set ``special_tags`` to interfaces.
-- Added endpoint ``DELETE v3/links/{link_id}`` to delete a disabled link.
+- Added endpoint ``DELETE v3/links/{link_id}`` to delete a disabled link. To avoid rediscovery, the link needs to be physically disconnected or both interface ends (switches) need to be disabled.
+- Added endpoint ``DELETE /v3/switches/{dpid}`` to delete a disabled switch. This endpoint is for decommissioned switches. The switch will get rediscovered if it reconnects to the controller. So, after deleting a switch on ``topology``, you're expected to also remove the TCP connection configuration on the switch.
 
 Deprecated
 ==========
@@ -29,12 +30,16 @@ Removed
 
 Fixed
 =====
+- An interface cannot be enabled if its switch is disabled.
 
 Security
 ========
 
 Changed
 =======
+- An interface can't be enabled if its switch is disabled.
+- A link can't be enabled if its interface is disabled.
+- Enabling interface can't not longer enabled its link. To enable a link, the endpoint ``POST v3/links/{link_id}/enable`` should be used.
 - If a KytosEvent can't be put on ``buffers.app`` during ``setup()``, it'll make the NApp to fail to start
 
 General Information